[incubator-kie-issues#1206] Removing unused methods/classes
