+++ kie-dmn/kie-dmn-feel/src/main/resources/TemplateCompiledFEELExpression.java
@@ -18,11 +18,8 @@
  */
 package org.kie.dmn.feel.codegen.feel11;
 
-import static org.kie.dmn.feel.codegen.feel11.CompiledFEELSemanticMappings.*;
-
 import org.kie.dmn.feel.codegen.feel11.CompiledCustomFEELFunction;
 import org.kie.dmn.feel.codegen.feel11.CompiledFEELExpression;
-import org.kie.dmn.feel.codegen.feel11.CompiledFEELSupport;
 import org.kie.dmn.feel.lang.ast.BaseNode;
 import org.kie.dmn.feel.lang.EvaluationContext;
 
@@ -35,7 +32,7 @@ public class TemplateCompiledFEELExpression implements org.kie.dmn.feel.codegen.
         try {
             return getBaseNode().evaluate(feelExprCtx);
         } catch (IllegalStateException e) {
-            return CompiledFEELSupport.notifyCompilationError(feelExprCtx, e.getMessage());
+            return notifyCompilationError(feelExprCtx, e.getMessage());
         }
     }
